# doc/openapi.yaml
swagger: '2.0'
info:
  contact: {}
  title: Breezy API
  description: API agrégée de la plateforme Breezy (réseau social microservices).
  version: '1.0'
paths:
...
  /users:
    get:
      produces:
      - application/json
      tags:
      - users
      summary: Lister les utilisateurs (paginé)
      parameters:
      - type: integer
        description: Nb résultats (défaut 20, max 100)
        name: limit
        in: query
      - type: integer
        description: Décalage
        name: offset
        in: query
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/models.User'
        '500':
          description: Internal Server Error
          schema:
            type: object
            additionalProperties:
              type: string
    post:
      security:
      - BearerAuth: []
      consumes:
      - application/json
      produces:
      - application/json
      tags:
      - users
      summary: Créer l'entrée utilisateur
      parameters:
      - description: Username
        name: body
        in: body
        required: true
        schema:
          $ref: '#/definitions/models.CreateUserRequest'
      responses:
        '201':
          description: Utilisateur créé
          schema:
            $ref: '#/definitions/models.User'
        '400':
          description: Payload invalide
          schema:
            type: object
            additionalProperties:
              type: string
        '401':
          description: Non authentifié
          schema:
            type: object
            additionalProperties:
              type: string
        '409':
          description: Username pris
          schema:
            type: object
            additionalProperties:
              type: string
  /users/by-username/{username}:
    get:
      produces:
      - application/json
      tags:
      - users
      summary: Détail d'un utilisateur par username
      parameters:
      - type: string
        description: Username (handle)
        name: username
        in: path
        required: true
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/models.UserDetails'
        '404':
          description: Not Found
          schema:
            type: object
            additionalProperties:
              type: string
...
  /users/me:
    get:
      security:
      - BearerAuth: []
      produces:
      - application/json
      tags:
      - users
      summary: Utilisateur courant (provisioning paresseux)
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        '401':
          description: Unauthorized
          schema:
            type: object
            additionalProperties:
              type: string
        '500':
          description: Internal Server Error
          schema:
            type: object
            additionalProperties:
              type: string
    patch:
      security:
      - BearerAuth: []
      consumes:
      - application/json
      produces:
      - application/json
      tags:
      - users
      summary: Modifier le username courant
      parameters:
      - description: Champs à modifier
        name: body
        in: body
        required: true
        schema:
          $ref: '#/definitions/models.UpdateUserRequest'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        '400':
          description: Bad Request
          schema:
            type: object
            additionalProperties:
              type: string
        '401':
          description: Unauthorized
          schema:
            type: object
            additionalProperties:
              type: string
        '409':
          description: Username pris
          schema:
            type: object
            additionalProperties:
              type: string
        '429':
          description: Cooldown actif
          schema:
            type: object
            additionalProperties:
              type: string
...
  /users/search:
    get:
      produces:
      - application/json
      tags:
      - users
      summary: Rechercher des utilisateurs par username
      parameters:
      - type: string
        description: Terme de recherche
        name: q
        in: query
        required: true
      - type: integer
        description: Nb résultats
        name: limit
        in: query
      - type: integer
        description: Décalage
        name: offset
        in: query
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/models.User'
        '500':
          description: Internal Server Error
          schema:
            type: object
            additionalProperties:
              type: string
  /users/suggestions:
    get:
      produces:
      - application/json
      tags:
      - users
      summary: Suggestions (comptes les plus suivis)
      parameters:
      - type: integer
        description: Nb résultats
        name: limit
        in: query
      - type: integer
        description: Décalage
        name: offset
        in: query
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/models.User'
        '500':
          description: Internal Server Error
          schema:
            type: object
            additionalProperties:
              type: string
  /users/{id}:
    get:
      produces:
      - application/json
      tags:
      - users
      summary: Détail d'un utilisateur par ID
      parameters:
      - type: string
        description: User ID
        name: id
        in: path
        required: true
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/models.UserDetails'
        '404':
          description: Not Found
          schema:
            type: object
            additionalProperties:
              type: string
    delete:
      security:
      - BearerAuth: []
      produces:
      - application/json
      tags:
      - users
      summary: Désactiver un compte (admin)
      parameters:
      - type: string
        description: User ID
        name: id
        in: path
        required: true
      responses:
        '204':
          description: No Content
        '401':
          description: Unauthorized
          schema:
            type: object
            additionalProperties:
              type: string
        '403':
          description: Réservé admin
          schema:
            type: object
            additionalProperties:
              type: string
        '404':
          description: Not Found
          schema:
            type: object
            additionalProperties:
              type: string
...
definitions:
...
  models.CreateUserRequest:
    type: object
    required:
    - username
    properties:
      username:
        type: string
        maxLength: 50
        minLength: 3
  models.UpdateUserRequest:
    type: object
    properties:
      username:
        type: string
        maxLength: 50
        minLength: 3
  models.User:
    type: object
    properties:
      created_at:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      updated_at:
        type: string
      username:
        type: string
      username_changed_at:
        description: 'UsernameChangedAt : date du dernier changement effectif du handle

          (nil = jamais changé). Sert de base au cooldown de changement de

          username (cf. service). Symétrique de display_name_changed_at (profil).'
        type: string
  models.UserDetails:
    type: object
    properties:
      created_at:
        type: string
      follower_count:
        type: integer
      following_count:
        type: integer
      id:
        type: string
      is_active:
        type: boolean
      updated_at:
        type: string
      username:
        type: string
      username_changed_at:
        description: 'UsernameChangedAt : date du dernier changement effectif du handle

          (nil = jamais changé). Sert de base au cooldown de changement de

          username (cf. service). Symétrique de display_name_changed_at (profil).'
        type: string
...
host: localhost:8080
basePath: /